create-bosh-lite script creates a sourceme with bosh creds
We also removed the bbl-state.json from the create-bosh-lite script. The
intent was to allow us to use `bbl print-env` to talk to bosh, but `bbl
print-env` requires the gcp auth json which we don't have on the vm.
Instead, we write a file that can be sourced for those creds.

Signed-off-by: Nitya Dhanushkodi <ndhanushkodi@pivotal.io>